Rbac

+++ pages/api/department_review/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import departmentReviewAbility from "@/services/ability/departmentReviewAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -20,7 +17,6 @@ export default async function handler(
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = departmentReviewAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -34,7 +30,6 @@ export default async function handler(
       const departmentReview = await prisma.departmentReview.findFirstOrThrow({
         where: {
           AND: [
-            accessibleBy(ability).DepartmentReview,
             {
               id: {
                 equals: id,
@@ -60,10 +55,16 @@ export default async function handler(
       });
 
       await validator.validate(req.query);
-
-      ForbiddenError.from(ability).throwUnlessCan("delete", "DepartmentReview");
-
       const { id } = validator.cast(req.query);
+
+      // ForbiddenError.from(ability).throwUnlessCan("delete", "DepartmentReview");
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const departmentReview = await prisma.departmentReview.delete({
         where: {
