+++ pages/api/qamis_dean_endorsement/[id]/index.tsx
@@ -1,5 +1,4 @@
 import prisma from "@/prisma/client";
-import QAMISDeanEndorsementAbility from "@/services/ability/qAMISDeanEndorsementAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
 import { ForbiddenError } from "@casl/ability";
@@ -20,7 +19,6 @@ export default async function handler(
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = QAMISDeanEndorsementAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -31,18 +29,18 @@ export default async function handler(
       await validator.validate(req.query);
 
       const { id } = validator.cast(req.query);
-      const QAMISDeanEndorsement = await prisma.qAMISDeanEndorsement.findFirstOrThrow({
-        where: {
-          AND: [
-            accessibleBy(ability).QAMISDeanEndorsement,
-            {
-              id: {
-                equals: id,
+      const QAMISDeanEndorsement =
+        await prisma.qAMISDeanEndorsement.findFirstOrThrow({
+          where: {
+            AND: [
+              {
+                id: {
+                  equals: id,
+                },
               },
-            },
-          ],
-        },
-      });
+            ],
+          },
+        });
 
       return res.json(QAMISDeanEndorsement);
     } catch (error: any) {
@@ -61,9 +59,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan("delete", "QAMISDeanEndorsement");
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const QAMISDeanEndorsement = await prisma.qAMISDeanEndorsement.delete({
         where: {
@@ -79,7 +83,6 @@ export default async function handler(
         .json({ error: { message: error?.message ?? "Server Error" } });
     }
   };
-
 
   switch (req.method) {
     case "DELETE":
