+++ pages/api/submitted_coordinator_suggestion/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import submittedCoordinatorSuggestionAbility from "@/services/ability/submittedCoordinatorSuggestionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -16,12 +13,10 @@ export default async function handler(
 
   try {
     user = await getServerUser(req, res);
-    981;
   } catch (error) {
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = submittedCoordinatorSuggestionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -36,7 +31,6 @@ export default async function handler(
         await prisma.submittedCoordinatorSuggestion.findFirstOrThrow({
           where: {
             AND: [
-              accessibleBy(ability).SubmittedCoordinatorSuggestion,
               {
                 id: {
                   equals: id,
@@ -62,13 +56,16 @@ export default async function handler(
       });
 
       await validator.validate(req.query);
-
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "SubmittedCoordinatorSuggestion"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(400).json({
+          error: {
+            message: "You are not allowed to perform this action"
+          }
+        })
+      }
+
       const submittedCoordinatorSuggestion =
         await prisma.submittedCoordinatorSuggestion.delete({
           where: {
