+++ pages/api/submitted_idd_specialist_suggestion/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import submittedIDDSpecialistSuggestionAbility from "@/services/ability/submittedIDDSpecialistSuggestionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -16,12 +13,10 @@ export default async function handler(
 
   try {
     user = await getServerUser(req, res);
-    981;
   } catch (error) {
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = submittedIDDSpecialistSuggestionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -37,7 +32,6 @@ export default async function handler(
         await prisma.submittedIDDSpecialistSuggestion.findFirstOrThrow({
           where: {
             AND: [
-              accessibleBy(ability).SubmittedIDDSpecialistSuggestion,
               {
                 id: {
                   equals: id,
@@ -64,12 +58,32 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "SubmittedIDDSpecialistSuggestion"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
+      
+      // const iMERCCITLReviewed = await prisma.iMERCCITLReviewed.findFirst({
+      //   where: {
+      //     SubmittedIDDSpecialistSuggestion: {
+      //       id: {
+      //         equals: id, 
+      //       },
+      //     },
+      //   },
+      // });
+      // if (iMERCCITLReviewed) {
+      //   await prisma.iMERCCITLReviewed.delete({
+      //     where: {
+      //       id: iMERCCITLReviewed.id,
+      //     },
+      //   });
+      // }
 
       const submittedIDDSpecialistSuggestion =
         await prisma.submittedIDDSpecialistSuggestion.delete({
