+++ pages/api/submitted_qamis_suggestion/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import submittedQAMISSuggestionAbility from "@/services/ability/submittedQAMISSuggestionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -16,12 +13,11 @@ export default async function handler(
 
   try {
     user = await getServerUser(req, res);
-    981;
+    
   } catch (error) {
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = submittedQAMISSuggestionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -36,7 +32,6 @@ export default async function handler(
         await prisma.submittedQAMISSuggestion.findFirstOrThrow({
           where: {
             AND: [
-              accessibleBy(ability).SubmittedQAMISSuggestion,
               {
                 id: {
                   equals: id,
@@ -63,12 +58,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "SubmittedQAMISSuggestion"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if(!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action"
+          }
+        })
+      }
       const submittedQAMISSuggestion =
         await prisma.submittedQAMISSuggestion.delete({
           where: {
